# [WIP] Significantly improve performance of Utf8String / Utf8Span GetCharCount method